docker/kafka-tools/requirements.txt
# Kafka clients
kafka-python==2.0.2
confluent-kafka==2.3.0

# Data processing
pandas==2.1.1
pyarrow==14.0.1

# AWS integration
boto3==1.34.0
s3fs==2023.12.0

# Data generation
faker==19.6.2

# Monitoring
prometheus-client==0.19.0